LayoutTests/imported/w3c/web-platform-tests/eventsource/request-cache-control.any-expected.txt

Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control due to access control checks.

Harness Error (TIMEOUT), message = null

PASS resources/cache-control.event_stream?pipe=sub1
PASS resources/cache-control.event_stream?pipe=sub2
TIMEOUT http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control1 Test timed out
NOTRUN http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control2
